IB/core: Change lid size in struct ib_wc from 16 to 32 bits
OPA devices can support larger number of nodes than 16 bit lids
can accommodate. This change increases the lid size from 16 to
32 bits in the kernel so that the full lid can be obtained in
the work completion.
UAPI is unchanged and is still only 16 bits.

+++ drivers/infiniband/core/uverbs_cmd.c
@@ -1593,7 +1593,7 @@ static int copy_wc_to_user(void __user *dest, struct ib_wc *wc)
 	tmp.src_qp		= wc->src_qp;
 	tmp.wc_flags		= wc->wc_flags;
 	tmp.pkey_index		= wc->pkey_index;
-	tmp.slid		= wc->slid;
+	tmp.slid		= (u16)wc->slid;
 	tmp.sl			= wc->sl;
 	tmp.dlid_path_bits	= wc->dlid_path_bits;
 	tmp.port_num		= wc->port_num;
